_WIP/requirements_GPU_pip.txt
numpy==1.16.4
opencv-python
requests
tensorflow-gpu==2.0.0-beta1
tqdm


